requirements.txt
flask
flask_bootstrap
flask_wtf
python-ldap
